v2/requirements.txt
alpaca-py
numpy
yfinance
matplotlib
statsmodels
ccxt
pandas
pyyaml
pyarrow
duckdb
pandas_market_calendars
python-dotenv
tqdm
ibapi